+++ pkg/lint/rules/examples_exist_test.go
@@ -13,18 +13,28 @@ func TestExampleExists(t *testing.T) {
 		nViolations int
 	}{
 		{
-			name:        "has no example",
+			name:        "restricted string and no examples",
 			schemaPath:  "testdata/examples_exist/no_examples.json",
-			nViolations: 1,
+			nViolations: 2,
 		},
 		{
-			name:        "has example",
+			name:        "restricted string and examples",
 			schemaPath:  "testdata/examples_exist/has_examples.json",
 			nViolations: 0,
 		},
 		{
 			name:        "no examples, but boolean",
 			schemaPath:  "testdata/examples_exist/no_examples_but_boolean.json",
+			nViolations: 0,
+		},
+		{
+			name:        "no examples, but integer",
+			schemaPath:  "testdata/examples_exist/no_examples_but_integer.json",
+			nViolations: 0,
+		},
+		{
+			name:        "no examples and string but not restricted",
+			schemaPath:  "testdata/examples_exist/no_examples_and_string_but_not_restricted.json",
 			nViolations: 0,
 		},
 	}
@@ -39,7 +49,12 @@ func TestExampleExists(t *testing.T) {
 			ruleResults := exampleExistsRule.Verify(schema)
 
 			if len(ruleResults.Violations) != tc.nViolations {
-				t.Fatalf("Unexpected number of rule violations in test case '%s': Expected %d, got %d", tc.name, tc.nViolations, len(ruleResults.Violations))
+				t.Fatalf(
+					"Unexpected number of rule violations in test case '%s': Expected %d, got %d",
+					tc.name,
+					tc.nViolations,
+					len(ruleResults.Violations),
+				)
 			}
 		})
 	}